refactor(sync): split coordinator contract from sqlite adapter
